Changing IProblem interface to return a long (Int64) Adding problem 8

+++ eulerproject/IProblem.cs
@@ -4,7 +4,7 @@ namespace eulerproject
 {
 	public interface IProblem
 	{
-		int Run ();
+		long Run ();
 	}
 }
